LayoutTests/media/modern-media-controls/fullscreen-support/fullscreen-support-enabled.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<style type="text/css" media="screen">
    
    video, #host {
        position: absolute;
        top: 0;
        left: 0;
    }

    video {
        width: 800px;
        height: 240px;
    }
    
</style>
<video src="../../content/test.mp4"></video>
<div id="host"></div>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>FullscreenSupport</code> behavior.");

const container = document.querySelector("div#host");
const media = document.querySelector("video");
const mediaController = createControls(container, media, null);

debug("Once media loads and we have tracks in the video, the fullscreen button should become enabled");
shouldBecomeEqual("mediaController.controls.fullscreenButton.enabled", "true", () => {
    media.addEventListener("error", () => {
        shouldBeFalse("mediaController.controls.fullscreenButton.enabled");
        debug("");
        container.remove();
        media.remove();
        finishJSTest();
    });
    debug("");
    debug("We now set the media source to be the empty string to force an error, which should disable the fullscreen button");
    media.src = "";
});

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
