LayoutTests/media/modern-media-controls/fullscreen-support/fullscreen-support-enabled-expected.txt

Testing the FullscreenSupport behavior.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Once media loads and we have tracks in the video, the fullscreen button should become enabled
PASS mediaController.controls.fullscreenButton.enabled became true

We now set the media source to be the empty string to force an error, which should disable the fullscreen button
PASS mediaController.controls.fullscreenButton.enabled is false

PASS successfullyParsed is true

TEST COMPLETE
